requirements.txt
flask
gunicorn
google-ads
pandas
matplotlib
numpy
python-dotenv